Move hard backspace to behaviours file
